Move QUIC from explicit work item to possibility
The discussion on the WG indicated that not everyone agreed
that QUIC is the obvious candidate for "the next protocol
to incorporate".

This CL moves the work on additional API for other data
protocols from an explicit QUIC work item to being part of
the "data transfer" functions bullet.

Fixes #27 (maybe)